Updating to OpenSSL 3.1.0
Adding all missing files.
Removing files that should not have been added by "OpenSSL 3.1.0".
All generated files are stored in repository, not generated on compile time + The templates are not in repository (the way we already did with the .asm files).
Some of the removed files might not have been needed anymore even before upgrade to OpenSSL 3.
So Perl is still needed for the generation only, not for the build.

Source commit: 53cb910ec00828fe1b1eefa6abf04fac3a4110db